Small changes, rename vector package to math
