+++ src/main/kotlin/game/GameManager.kt
@@ -2,8 +2,8 @@ package asteroid4.tileengine.game
 
 import asteroid4.tileengine.ProgramData
 import asteroid4.tileengine.Registries
-import asteroid4.tileengine.game.vector.FloatVector
-import asteroid4.tileengine.game.vector.IntVector
+import asteroid4.tileengine.game.math.FloatVector
+import asteroid4.tileengine.game.math.IntVector
 import asteroid4.tileengine.game.world.Tile
 import asteroid4.tileengine.game.world.World
 import java.awt.Image
@@ -14,22 +14,24 @@ class GameManager() {
 
     init {
         ProgramData.INPUT_MANAGER.register('w') {
-            velocity += FloatVector(0f, 1f)
+            velocity += FloatVector(0f, 0.01f)
         }
         ProgramData.INPUT_MANAGER.register('a') {
-            velocity += FloatVector(-1f, 0f)
+            velocity += FloatVector(-0.01f, 0f)
         }
         ProgramData.INPUT_MANAGER.register('s') {
-            velocity += FloatVector(0f, -1f)
+            velocity += FloatVector(0f, -0.01f)
         }
         ProgramData.INPUT_MANAGER.register('d') {
-            velocity += FloatVector(1f, 0f)
+            velocity += FloatVector(0.01f, 0f)
         }
     }
 
     fun tick() {
         if (currentWorld == null) return
+        ProgramData.LOGGER.print(velocity.toString())
         ProgramData.LOGGER.print(currentWorld!!.player.position.toString())
+        ProgramData.LOGGER.newLine()
         currentWorld!!.player.position += velocity
     }
