+++ src/main/kotlin/game/world/entity/ControllableEntity.kt
@@ -1,5 +1,5 @@
 package asteroid4.tileengine.game.world.entity
 
-import asteroid4.tileengine.game.vector.FloatVector
+import asteroid4.tileengine.game.math.FloatVector
 
 class ControllableEntity(position: FloatVector) : Entity(position)
